chore: add tsdown helper
