+++ packages/workbox-build/tsdown.config.ts
@@ -1,13 +1,9 @@
 import { defineConfig } from 'tsdown'
+import { banner, exports } from '../../helper.ts'
 
 export default defineConfig({
   entry: './src/{index,types,generate-sw,get-manifest,inject-manifest}.ts',
   platform: 'node',
-  banner: `/*
-  Copyright 2019 Google LLC, Vite PWA's Team
-
-  Use of this source code is governed by an MIT-style
-  license that can be found in the LICENSE file or at
-  https://opensource.org/licenses/MIT.
-*/`,
+  banner,
+  exports,
 })
