+++ tsconfig.json
@@ -4,6 +4,7 @@
     "lib": ["DOM", "ESNext", "DOM.Iterable", "ESNext.AsyncIterable", "WebWorker"],
     "module": "ESNext",
     "moduleResolution": "Bundler",
+    "allowImportingTsExtensions": true,
     "paths": {
       "@vite-pwa/workbox-build": [
         "./packages/workbox-build/src/index.ts"
